+++ src/classes/screens/game-screen.js
@@ -8,9 +8,62 @@ export class GameScreen extends Screen {
    * @param background
    * @param showWithMe[]
    */
-  constructor({selector, triggers, backgroundSelector, showWithMe}) {
+  constructor({selector, triggers, backgroundSelector, showWithMe, hasFrame, frameID}) {
     super({selector, triggers, backgroundSelector, showWithMe})
+    this._hasFrame = hasFrame ?? false
+    this._frameID = frameID ?? null
     this._ctx = this._htmlEL.getContext("2d")
+  }
+
+  displayFrame() {
+    if (this._hasFrame) {
+      console.log('displaying frame', this._frameID)
+      if (this._frameID) {
+        let gameFrame = document.getElementById(this._frameID)
+        if (gameFrame) {
+          gameFrame.style.display = 'block'
+        } else {
+          throw new Error(`gameFrame with id ${this._frameID} not found`)
+        }
+      } else {
+        throw new Error('frameID not set')
+      }
+    }
+  }
+
+  updateFrame() {
+    if (this._hasFrame) {
+      if (this._frameID) {
+        let gameFrame = document.getElementById(this._frameID)
+        if (gameFrame) {
+          const gameCanvas = document.getElementById('my-canvas')
+          const canvasRect = gameCanvas.getBoundingClientRect();
+          gameFrame.style.top = `${canvasRect.top - 1166}px`;
+          gameFrame.style.left = `${canvasRect.left - 81}px`;
+          gameFrame.style.width = `${canvasRect.width + 314}px`;
+          gameFrame.style.height = `${canvasRect.height + 1260}px`;
+        } else {
+          throw new Error(`gameFrame with id ${this._frameID} not found`)
+        }
+      } else {
+        throw new Error('frameID not set')
+      }
+    }
+  }
+
+  hideFrame() {
+    if (this._hasFrame) {
+      if (this._frameID) {
+        let gameFrame = document.getElementById(this._frameID)
+        if (gameFrame) {
+          gameFrame.style.display = 'none'
+        } else {
+          throw new Error(`gameFrame with id ${this._frameID} not found`)
+        }
+      } else {
+        throw new Error('frameID not set')
+      }
+    }
   }
 
   getCanvas() {
